server refactor
