+++ backend/server.js
@@ -1,40 +1,10 @@
 import express from 'express'
 import bodyParser from 'body-parser'
 import cors from 'cors'
-import mongoose from 'mongoose'
-import dotenv from 'dotenv'
-import crypto from 'crypto'
-import bcrypt from 'bcrypt-nodejs'
 
-dotenv.config()
+import usersRoutes from './routes/users'
+import sessionsRoutes from './routes/sessions'
 
-const mongoUrl = process.env.MONGO_URL || "mongodb://localhost/authAPI"
-mongoose.connect(mongoUrl, { useNewUrlParser: true, useUnifiedTopology: true })
-mongoose.Promise = Promise
-
-const User = mongoose.model('User', {
-  name: {
-    type: String,
-    unique: true
-  },
-  email: {
-    type: String,
-    unique: true
-  },
-  password: {
-    type: String,
-    required: true
-  },
-  accessToken: {
-    type: String,
-    default: () => crypto.randomBytes(128).toString('hex')
-  }
-})
-
-// Defines the port the app will run on. Defaults to 8080, but can be 
-// overridden when starting the server. For example:
-//
-//   PORT=9000 npm start
 const port = process.env.PORT || 8080
 const app = express()
 
@@ -42,59 +12,13 @@ const app = express()
 app.use(cors())
 app.use(bodyParser.json())
 
-const authenticateUser = async (req, res, next) => {
-  try {
-    const user = await User.findOne({
-      accessToken: req.header('Authorization')
-    })
-    if (user) {
-      req.user = user
-      next()
-    } else {
-      res.status(401).json({ loggedOut: true })
-    }
-  } catch (err) {
-    res
-      .status(403)
-      .json({ message: 'access token missing or wrong', errors: err.errors })
-  }
-}
-
 // Start defining your routes here
 app.get('/', (req, res) => {
   res.send('Hello world')
 })
+app.use('/users', usersRoutes)
 
-app.post('/users', async (req, res) => {
-  try {
-    const { name, email, password } = req.body
-    const user = new User({ name, email, password: bcrypt.hashSync(password) })
-    const saved = await user.save()
-    res.status(201).json({ saved })
-  } catch (err) {
-    res.status(400).json({ message: 'could not create user', error: err.errors })
-  }
-})
-
-app.get('/users/:id', authenticateUser)
-app.get('/users/:id', async (req, res) => {
-  const user = await User.findById(req.params.id)
-  res.json({ secret: `Welcome to the Jungle!! ${user.name}` })
-})
-
-app.post('/sessions', async (req, res) => {
-  try {
-    const { email, password } = req.body
-    const user = await User.findOne({ email })
-    if (user && bcrypt.compareSync(password, user.password)) {
-      res.status(201).json({ userId: user._id, accessToken: user.accessToken })
-    } else {
-      throw new Error('could not find user')
-    }
-  } catch (err) {
-    res.status(400).json({ message: 'could not find user', errors: err.errors })
-  }
-})
+app.use('/sessions', sessionsRoutes)
 
 // Start the server
 app.listen(port, () => {
